remove unnecessary headers. move Game into game.hpp. start unpacking move keys. move stuff around
