+++ src/util.cpp
@@ -14,6 +14,15 @@ uint32_t generate_move_key(uint8_t src_square, uint8_t dest_square, uint8_t prom
     return (src_square << 16) + (dest_square << 8) + promotion_piece;
 }
 
+Move unpack_move_key(uint32_t move_key)
+{
+    Move move;
+    move.src_square = move_key >> 16;
+    move.dest_square = (move_key & 0xff00) >> 8;
+    move.promotion_piece = move_key & 0xff;
+    return move;
+}
+
 char getc(int i, std::smatch &matches)
 {
     if (matches[i].length())
